Chore: moving openvasd into new style
