+++ rust/src/openvasd/container_image_scanner/scheduling/db/mod.rs
@@ -1,2 +1,2 @@
-mod sqlx;
-pub use sqlx::*;
+mod sqlite;
+pub use sqlite::*;
