+++ rust/src/openvasd/container_image_scanner/scheduling/db/sqlite/preferences.rs
