+++ rust/src/openvasd/database/sqlite/results.rs
